+++ src/main/java/io/openliberty/tools/intellij/lsp4jakarta/lsp4ij/interceptor/Constants.java
@@ -23,6 +23,7 @@ public class Constants {
     /* INTERCEPTOR_FQ_NAME */
     public static final String INTERCEPTOR_FQ_NAME = "jakarta.interceptor.Interceptor";
     public static final String INTERCEPTOR_BINDING_FQ_NAME = "jakarta.interceptor.InterceptorBinding";
+    public static final String INTERCEPTORS_FQ_NAME = "jakarta.interceptor.Interceptors";
     /* Source */
     public static final String DIAGNOSTIC_SOURCE = "jakarta-interceptor";
     public static final String DIAGNOSTIC_CODE_INTERCEPTOR_ON_ABSTRACT_CLASS = "RemoveInterceptorAnnotationOnAbstractClass";
@@ -36,8 +37,6 @@ public class Constants {
     public static final String DIAGNOSTIC_CODE_MISSING_INTERCEPTOR_BINDING = "InvalidInterceptorMissingInterceptorBinding";
     public static final String DIAGNOSTIC_CODE_FINAL_INTERCEPTOR_BINDING_CLASS = "InvalidFinalInterceptorBindingClass";
     public static final String DIAGNOSTIC_CODE_FINAL_METHOD_ON_INTERCEPTOR_BINDING_CLASS = "InvalidMethodOnInterceptorBindingClass";
-
-    public static final String INTERCEPTORS_FQ_NAME = "jakarta.interceptor.Interceptors";
 
     private static final String AROUND_CONSTRUCT_FQ_NAME = "jakarta.interceptor.AroundConstruct";
