+++ src/test/java/io/openliberty/tools/intellij/lsp4jakarta/it/interceptor/JakartaInterceptorTest.java
@@ -1551,29 +1551,45 @@ public class JakartaInterceptorTest extends BaseJakartaTest {
         JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
         diagnosticsParams.setUris(Arrays.asList(uri));
 
-        // Line 29: final class InvalidFinalInterceptorBindingClass {
-        Diagnostic finalClassDiagnostic = JakartaForJavaAssert.d(28, 12, 47,
+        // Line 27: final class InvalidFinalInterceptorBindingClass {
+        Diagnostic finalClassDiagnostic = JakartaForJavaAssert.d(26, 12, 47,
                 "A component class that declares or inherits a class-level interceptor binding must not be declared final.",
                 DiagnosticSeverity.Error, "jakarta-interceptor", "InvalidFinalInterceptorBindingClass");
 
-        // Line 43: public final Object intercept(...) - non-static, non-private final method
-        Diagnostic finalInterceptDiagnostic = JakartaForJavaAssert.d(42, 24, 33,
+        // Line 40: public final Object intercept(...) - non-static, non-private final method
+        Diagnostic finalInterceptDiagnostic = JakartaForJavaAssert.d(39, 24, 33,
                 "A component class that declares or inherits a class-level interceptor binding must not have a non-static, non-private final method 'intercept'.",
                 DiagnosticSeverity.Error, "jakarta-interceptor", "InvalidMethodOnInterceptorBindingClass");
 
-        // Line 48: protected final void protectedHelper() - non-static, non-private final method
-        Diagnostic finalProtectedHelperDiagnostic = JakartaForJavaAssert.d(47, 25, 40,
+        // Line 45: protected final void protectedHelper() - non-static, non-private final method
+        Diagnostic finalProtectedHelperDiagnostic = JakartaForJavaAssert.d(44, 25, 40,
                 "A component class that declares or inherits a class-level interceptor binding must not have a non-static, non-private final method 'protectedHelper'.",
                 DiagnosticSeverity.Error, "jakarta-interceptor", "InvalidMethodOnInterceptorBindingClass");
 
-        // Line 52: final void packagePrivateHelper() - package-private (default) final method, non-static, non-private
-        Diagnostic finalPackagePrivateHelperDiagnostic = JakartaForJavaAssert.d(51, 15, 35,
+        // Line 49: final void packagePrivateHelper() - package-private (default) final method, non-static, non-private
+        Diagnostic finalPackagePrivateHelperDiagnostic = JakartaForJavaAssert.d(48, 15, 35,
                 "A component class that declares or inherits a class-level interceptor binding must not have a non-static, non-private final method 'packagePrivateHelper'.",
                 DiagnosticSeverity.Error, "jakarta-interceptor", "InvalidMethodOnInterceptorBindingClass");
 
         JakartaForJavaAssert.assertJavaDiagnostics(diagnosticsParams, utils,
                 finalClassDiagnostic, finalInterceptDiagnostic,
                 finalProtectedHelperDiagnostic, finalPackagePrivateHelperDiagnostic);
+    }
+
+    @Test
+    public void testValidInterceptorBindingClassModifiers() throws Exception {
+        Module module = createMavenModule(new File("src/test/resources/projects/maven/jakarta-sample"));
+        IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
+
+        VirtualFile javaFile = LocalFileSystem.getInstance().refreshAndFindFileByPath(ModuleUtilCore.getModuleDirPath(module)
+                + "/src/main/java/io/openliberty/sample/jakarta/interceptor/ValidInterceptorBindingClassModifiers.java");
+        String uri = VfsUtilCore.virtualToIoFile(javaFile).toURI().toString();
+
+        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
+        diagnosticsParams.setUris(Arrays.asList(uri));
+
+        // Test that valid class and methods with interceptor binding do NOT trigger any diagnostic
+        JakartaForJavaAssert.assertJavaDiagnostics(diagnosticsParams, utils);
     }
 
     @Test
