+++ src/test/resources/projects/maven/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/interceptor/ValidInterceptorBindingClassModifiers.java
@@ -1,0 +1,38 @@
+/*******************************************************************************
+ * Copyright (c) 2026 IBM Corporation and others.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial implementation
+ *******************************************************************************/
+package io.openliberty.sample.jakarta.interceptor;
+
+import jakarta.interceptor.AroundInvoke;
+import jakarta.interceptor.InvocationContext;
+
+/**
+ * Valid dataset for issue #693:
+ * A component class that declares or inherits a class-level interceptor binding
+ * (@Monitored) with no final class modifier and valid methods.
+ */
+@Monitored
+public class ValidInterceptorBindingClassModifiers {
+
+    @AroundInvoke
+    public Object intercept(InvocationContext ctx) throws Exception {
+        return ctx.proceed();
+    }
+
+    // Valid: public non-final method
+    public void publicHelper() {
+    }
+
+    // Valid: public static non-final method
+    public static void publicStaticNonFinalHelper() {
+    }
+}
